% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predicted_values.R
\name{calculate_predicted_values}
\alias{calculate_predicted_values}
\title{Calculates Predicted Values}
\usage{
calculate_predicted_values(
  model,
  type = "link",
  newdata = NULL,
  add_input = FALSE
)
}
\arguments{
\item{model, }{object of type lmerMod or clmm}

\item{type, }{whether to use link or probabilites}

\item{add_input, }{logical}

\item{newata, }{must be a data.frame with matching variable names}
}
\value{
a Dataframe
}
\description{
Calculates Predicted Values from lmerMod and clmm objects
}
