% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeTools-package.R
\docType{package}
\name{crimeTools-package}
\alias{crimeTools}
\alias{crimeTools-package}
\title{crimeTools: Data Analysis and Visualization Tools for Crime Analysis}
\description{
crimeTools is a package designed to make analysis and visualization of crime survey data straightforward by providing high-level functions for the most common chart types. crimeTools builds on the R packages survey and srvyr to properly weight the results in accordance with the survey and sampling design.
}
\keyword{internal}
