% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Returns a density or ridge density plot}
\usage{
plot_density(
  data,
  x,
  group = NULL,
  area_limits = NULL,
  ridge_density = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  legend = "standard",
  legend_title = NULL,
  scale_x_breaks = waiver(),
  scale_x_labels = waiver(),
  scale_y_labels = waiver(),
  scale = 1.5,
  alpha = 1,
  fill = "#3e6487",
  area_fill = "#e36c33",
  group_color = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{x}{variable names, must be a symbol or character vector}

\item{group}{group (optional)}

\item{y}{variable names, must be a symbol or character vector}
}
\value{
A-ggplot()-object
}
\description{
Returns a density or ridge density plot
}
