% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_idw.R
\name{plot_idw}
\alias{plot_idw}
\title{Returns an interpolated map}
\usage{
plot_idw(
  data,
  location_data,
  formula,
  border_data = NULL,
  low = "#002664",
  high = "#F6D2C1",
  contour = TRUE,
  title = NULL,
  subtitle = NULL,
  legend_labels = scales::comma_format(accuracy = 0.1),
  ...
)
}
\arguments{
\item{data}{spatial dataframe of type sf}

\item{location_data}{spatial dataframe of type sf}

\item{formula}{formula string}
}
\value{
A-ggplot()-object
}
\description{
Returns an interpolated map using inverse distance weighting
}
