% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Plot Lollipop Charts}
\usage{
plot_lollipop(
  data,
  x,
  y,
  group = NULL,
  text,
  group_color = NULL,
  nudge_x = 0.05,
  alpha = 0.6,
  limits = NULL,
  xlab = NULL,
  ylab = NULL,
  position_dodge_width = 0.9,
  point_color = "#e36c33",
  point_size = 4,
  segment_color = "#3e6487",
  fontface = "plain",
  flip = FALSE,
  title = NULL,
  subtitle = NULL,
  legend = "standard",
  legend_size = 12,
  font_family = "",
  axis_text_size = 12,
  text_label_align = 0.5,
  text_label_size = 4,
  expand = expansion(mult = 0, add = 0),
  scale_x_labels = scales::percent_format(scale = 100, big.mark = ".", decimal.mark =
    ","),
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{x}{variable names, must be a symbol or character vector}

\item{y}{variable names, must be a symbol or character vector}

\item{group}{group (optional)}
}
\value{
A-ggplot()-object
}
\description{
Plots lollipop charts with or without grouping.
}
\examples{
plot_lollipop(data = iris, vars = setosa) ;
}
