% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mirt.R
\name{plot_mirt}
\alias{plot_mirt}
\title{Plots results of item reponse theory models}
\usage{
plot_mirt(
  data,
  itemtype = "gpcm",
  dim_name = "value",
  model,
  legend_categories = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{itemtype}{type of IRT model}

\item{dim_name}{name of dimension, must be a string}

\item{model}{model formula, must be a formula or string}

\item{legend_categories}{name of categories, must be a character vector}

\item{xlim}{limits of x axis of itemparameter plot, must be a numeric vector}

\item{ylim}{limits of y axis of itemparameter plot, must be a numeric vector}

\item{...}{further arguments for function mirt()}
}
\value{
A-irtClass()-object
}
\description{
Returns an object of class 'irtClass'
}
