% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prevalence.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Calculates and Plots Prevalence Values}
\usage{
plot_prevalence(
  data,
  vars,
  group = NULL,
  grouping = "group_by_vars",
  grouping_title = FALSE,
  prop_values,
  graph_type = "bar",
  na_values = NULL,
  na_group = NULL,
  show_n = TRUE,
  show_confint = TRUE,
  conf_level = 0.95,
  sort = "asc",
  group_sort_val = NULL,
  dec = 0,
  axis_text_width = 35,
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{group}{group (optional)}

\item{x}{variable names, must be a symbol or character vector}

\item{y}{variable names, must be a symbol or character vector}
}
\value{
A-ggplot()-object
}
\description{
plot_prevalence() is a wrapper around plot_bar() and plot_lollipop(). The function calculates prevalence values and their respective confidence intervals before plotting.
}
\examples{
plot_prevalence(data = iris, vars = setosa) ;
}
