% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_tidy.R
\name{wrap_tidy}
\alias{wrap_tidy}
\title{Wrapper around tidy()}
\usage{
wrap_tidy(model, conf.level = c(0.9, 0.95, 0.99))
}
\arguments{
\item{model}{Object of class lm, lmer or similar}

\item{conf.level}{confidence levels}
}
\value{
A-tibble()-object
}
\description{
Returns as tibble with multiple confidence intervals.
}
\examples{
wrap_tidy(model) ;
}
