% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeTools-package.R
\docType{package}
\name{crimeTools-package}
\alias{crimeTools}
\alias{crimeTools-package}
\title{crimeTools: Data Analysis and Visualization Tools for Crime Analysis}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
