% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likert.R
\name{plot_likert}
\alias{plot_likert}
\title{Plot Likert Scales as Centered Stacked Bars}
\usage{
plot_likert(
  data,
  vars,
  group = NULL,
  weight = NULL,
  reverse_coding = FALSE,
  bar_color_manual = NULL,
  bar_color_auto = c("#fce629", "#460456"),
  bar_width = 0.65,
  na_values = NULL,
  na_group = NULL,
  na_drop = TRUE,
  na_plot = "bar",
  na_bar_color = "grey30",
  na_bar_text_color = "grey",
  na_bar_text_nudge = 0.05,
  na_graph_nudge = 0.1,
  na_graph_expand = c(0.015, 0),
  na_graph_limits = c(-0.4, 1),
  na_label = "missing values",
  valid_label = "valid values",
  text_position = "inside",
  text_outside_nudge = 0.05,
  text_color = "black",
  text_middle_color = "black",
  text_label_size = 4,
  show_n = FALSE,
  sort = TRUE,
  sort_func = sum,
  sort_val = c(1, 2),
  dec = 0,
  axis_text_width = 35,
  legend = "standard",
  alpha = 1,
  guide = guide_legend(nrow = 1, override.aes = list(alpha = 1)),
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  font_family = "",
  ...
)
}
\arguments{
\item{data}{Data frame}
}
\value{
A-ggplot2::ggplot()-object or a list()
}
\description{
Plots likert scales with even and odd number of categories as centered stacked bar charts with and without missing values.
}
\examples{
plot_likert(data = iris, vars = setosa) ;
}
