% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likert.R
\name{plot_likert}
\alias{plot_likert}
\title{Plot Likert Scales as Centered Stacked Bars}
\usage{
plot_likert(
  data,
  vars,
  group = NULL,
  reverse_coding = FALSE,
  bar_color_manual = NULL,
  bar_color_auto = c("#fce629", "#460456"),
  bar_width = 0.65,
  na_values = NULL,
  na_group = NULL,
  na_drop = TRUE,
  na_plot = "bar",
  na_bar_color = "grey30",
  na_bar_text_color = "grey",
  na_bar_text_nudge = 0.05,
  na_graph_nudge = 0.1,
  na_graph_expand = c(0.015, 0),
  na_graph_limits = c(-0.4, 1),
  na_label = "missing values",
  valid_label = "valid values",
  text_position = "inside",
  text_outside_nudge = 0.05,
  text_color = "black",
  text_middle_color = "black",
  text_label_size = 4,
  show_n = FALSE,
  sort = TRUE,
  sort_func = sum,
  sort_val = c(1, 2),
  dec = 0,
  axis_text_width = 35,
  legend = "standard",
  alpha = 1,
  guide = guide_legend(nrow = 1, override.aes = list(alpha = 1)),
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  font_family = "",
  ...
)
}
\arguments{
\item{data}{data frame or a weighted data frame of class 'tbl_svy'.}

\item{vars}{variable names, must be a symbol or character vector.}

\item{group}{(optional) only used if grouping variable is supplied.}

\item{reverse_coding}{reverse coding of valid response categories.}

\item{bar_color_manual}{(optional) user defined colors, must be a character vector of same length as the number of valid response categories.}

\item{bar_color_auto}{define color space in RGB or CIE, must be a character vector of length 2.}

\item{na_values}{(optional) define missing values present in vars.}

\item{na_group}{(optional, only used if group != NULL) define missing values present in group.}

\item{na_drop}{logical, if TRUE missing values defined by na_values are dropped.}

\item{na_plot}{determines how missing values are display if na_drop = TRUE.
'bar' (default) displays missing values as separate bar
'graph' displays missing values as separate bar chart}

\item{na_bar_color}{character, defines bar color of na_plot}

\item{na_bar_text_color}{character, defines label color if na_plot = 'bar'}

\item{na_bar_text_nudge}{numeric, adjusts  label position if na_plot = 'bar'}
}
\value{
A-ggplot2::ggplot()-object or a list()
}
\description{
Plots likert scales with even and odd number of categories as centered stacked bar charts with and without missing values.
}
\examples{
library(sjmisc)
data(efc)

plot_likert(data = efc, vars = c82cop1:c90cop9)

# Plot different variable by group combinations
plot_likert(data = efc, vars = c82cop1:c90cop9, group = e16sex)

# use a weighted dataset
library(srvyr)
efc_weighted <- as_survey_design(.data = efc, ids = 1)

plot_likert(data = efc_weighted, vars = c82cop1:c90cop9)
}
