% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Returns a Coefficient Plot}
\usage{
plot_model(
  data,
  x,
  y,
  group = NULL,
  ymin_90,
  ymax_90,
  ymin_95,
  ymax_95,
  ymin_99,
  ymax_99,
  facet_group_row = NULL,
  patchwork_group = NULL,
  shape_values = NULL,
  shape_sizes = NULL,
  conf_colors = c("#3e6487", "#829cb2", "#c0cdd8"),
  strip_color = "#fbede6",
  patchwork_color = "#f6d2c1",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  breaks = 5,
  ylab = "Regressionskoeffizienten",
  xlab = "",
  xlabels,
  accuracy = 0.01,
  ylim = NULL,
  scales = "free_y",
  space = "free",
  guide_nrow = 3,
  guide_ncol = 2,
  legend_position = "bottom",
  panel_spacing = 0.5,
  font_family = "Arial",
  axis_text_size = 13,
  outer_text_size = 13,
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{x}{variable names, must be a symbol or character vector}

\item{y}{variable names, must be a symbol or character vector}

\item{group}{group (optional)}
}
\value{
A-ggplot()-object
}
\description{
Returns a Coefficient Plot
}
