% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardizes and normalizes numeric vectors.}
\usage{
standardize(variable, type = "divide_by_sd", times = 2)
}
\arguments{
\item{variable}{a numeric vector}

\item{type, }{takes the values "divide_by_sd" and "normalize"}
}
\value{
a numeric vector
}
\description{
Returns a vector of standardized values.
}
\examples{
x <- rnorm(100)
standardize(x) ;
}
